Extract magic strings into command.* constants
