+++ cmd/cluster-agent/subcommands/check/command.go
@@ -16,15 +16,13 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const loggerName = "CLUSTER"
-
 // Commands returns a slice of subcommands for the 'cluster-agent' command.
 func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 	cmd := check.MakeCommand(func() check.GlobalParams {
 		return check.GlobalParams{
 			ConfFilePath: globalParams.ConfFilePath,
 			ConfigName:   "datadog-cluster",
-			LoggerName:   loggerName,
+			LoggerName:   command.LoggerName,
 		}
 	})
